Move browser-only mqtt functionality to browser
